feat(SemestersFormList.jsx): cambiar el nombre de la materia por su código feat(Career.jsx): añadir ToastContainer de react-toastify feat(CareerContext.jsx): añadir renderToast para mostrar un mensaje de éxito al borrar la carrera feat(subjects.js): añadir una propiedad condition para indicar si una materia está bloqueada o desbloqueada fix(Career.js): añadir trim a los campos de texto para evitar espacios en blanco al principio o al final
feat([user].js): añadir ícono de estado a las condiciones de las materias y mostrar las materias del primer semestre de la carrera del usuario en la página de perfil
feat(api/career/delete.js): actualizar la lista de carreras del usuario al eliminar una carrera
fix(api/career/delete.js): manejar el caso en que no se proporcione un id en la solicitud
feat(career/create.js): mostrar un mensaje de éxito al crear una carrera
feat(settings/[option].js): agregar un enlace a la página de perfil del usuario en la sección de cuenta personal

+++ src/pages/api/career/delete.js
@@ -1,5 +1,6 @@
 import { dbConnect } from '@/utils/mongoose';
 import Career from '@/models/Career';
+import User from '@/models/User';
 
 const deleteCareer = async (req, res) => {
 
@@ -7,9 +8,25 @@ const deleteCareer = async (req, res) => {
 
     const { id } = req.query;
 
-    await Career.findByIdAndDelete(id);
+    if (id) {
+        try {
+            const user = await User.find({ careers: id });
+            if (user.length > 0) {
+                const updatedUser = await User.findByIdAndUpdate(user[0]._id, { careers: [] });
+            }
+        }
+        catch (err) {
+            console.log(err);
+            return res.status(500).json({ message: 'Error deleting career' });
+        }
+        finally {
+            await Career.findByIdAndDelete(id);
+            return res.json({ success: true, message: 'Career deleted' });
+        }
+    } else {
+        return res.status(400).json({ message: 'No id provided' });
+    }
 
-    res.json({ success: true, message: 'Career deleted' });
 }
